// src/modules/vsf-storyblok-module/types/spacing-setting-field-name.value.ts
export enum SpacingSettingsFieldName {
  MARGIN_TOP = 'margin_top',
  MARGIN_RIGHT = 'margin_right',
  MARGIN_BOTTOM = 'margin_bottom',
  MARGIN_LEFT = 'margin_left',
  PADDING_TOP = 'padding_top',
  PADDING_RIGHT = 'padding_right',
  PADDING_BOTTOM = 'padding_bottom',
  PADDING_LEFT = 'padding_left',
}